feat: reorganize styles for Logout and Register components into auth directory
