+++ src/styles/components/auth/Register.module.scss
@@ -1,4 +1,4 @@
-@use "../partials/variables" as *;
+@use "../../partials/variables" as *;
 .register-form-container {
   max-width: 40rem;
   font-family: $font-family;
